Project import generated by Copybara.
GitOrigin-RevId: 60c16c3db7a85947905bdcae04b49457512026a3
