+++ assemblyai/__init__.py
@@ -68,6 +68,7 @@ from .types import (
     SyncTranscriptError,
     SyncTranscriptionConfig,
     SyncTranscriptResponse,
+    SyncWord,
     Timestamp,
     TranscriptError,
     TranscriptionConfig,
@@ -148,6 +149,7 @@ __all__ = [
     "SyncTranscriptError",
     "SyncTranscriptionConfig",
     "SyncTranscriptResponse",
+    "SyncWord",
     "Timestamp",
     "Transcriber",
     "TranscriptionConfig",
