+++ assemblyai/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "0.64.26"
+__version__ = "0.64.29"
