+++ assemblyai/sync/v1/__init__.py
@@ -1,0 +1,19 @@
+from ...types import (
+    SyncSpeechModel,
+    SyncTranscriptError,
+    SyncTranscriptionConfig,
+    SyncTranscriptResponse,
+    SyncWord,
+)
+from ._base import AudioInput
+from .client import SyncTranscriber
+
+__all__ = [
+    "AudioInput",
+    "SyncSpeechModel",
+    "SyncTranscriber",
+    "SyncTranscriptError",
+    "SyncTranscriptionConfig",
+    "SyncTranscriptResponse",
+    "SyncWord",
+]
